codeclimate npm on dev build - improve Travis CI caching
As a separate dependencie installed only on Travis CI build, it's
taking separate time, as it not mention package version.
There is no point in specifying version on travis build config file
so the better move is to move the installation to package.json.

This way, we can do a proper versioning of this package and save some
seconds on build time.